chore: Add scaling job for Edify component
Add a new scaling job for the Edify component in the Exivity application. This scaling job is responsible for managing the parallel execution of Edify tasks. It includes volume mounts for configuration files, log files, and extracted data. The job runs a container with the Edify image and sets environment variables for backend log level and Prometheus push gateway. Additionally, it includes the necessary RabbitMQ authentication and sets the security context for the container. The job also exposes port 2112 for monitoring and includes a liveness probe. This change is necessary to improve the scalability and performance of the Edify component.

Signed-off-by: Steffen Exler <steffen@exivity.com>