chore: Update scaling job environment variables
This commit updates the scaling job environment variables in the `scaling-job.yaml` file for the Edify, Executor, Horizon, and Transcript components. It adds a new environment variable `LOG_LEVEL` with the value of `{{ .Values.logLevel.backend }}`. This change is necessary to align the environment variables across the components and improve consistency.

Signed-off-by: Steffen Exler <steffen@exivity.com>