chore: Update liveness probe configuration for components
This commit updates the liveness probe configuration for the Edify, Executor, Horizon, and Transcript components. It renames the existing `_liveness.tpl` template file to `_liveness_nfs.tpl` and adds a new template file `_liveness_tristan.tpl`. The new template file includes the necessary liveness probe configuration for the Tristan component. This change is necessary to improve the reliability and availability of the components.

Signed-off-by: Steffen Exler <steffen@exivity.com>